Revert "refactor: code review (#10)"
This reverts commit 8f4add31e4a12e688426f4bef806527dd6bf3448.
